+++ modules_3d/grid_set_up.py
@@ -1,4 +1,5 @@
-from modules_3d import board_set_up, error_flags
+from modules_3d import board_set_up
+from modules_gui import error_flags
 
 def create_grid(self):
     self.grid = [[0 for x in range(9)] for i in range(len(self.faces))] ## creating the grid
